Creación de pagina web y controller.producto
Se creo la pagina web usando bootstrap 3.7 se agregaron los archivos css
y js para funcionabilidad, Se creo un controlador usando angularJs

+++ MantenimientoProducto/src/main/java/com/cecylopez/MantenimientoProducto/entities/CategoriaProducto.java
@@ -4,14 +4,12 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.Table;
 
 @Entity
-
 public class CategoriaProducto {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	int codigo;
+	long codigo;
 	String nombre;
 	public long getCodigo() {
 		return codigo;
